+++ src/app/Modules/admin/Departments/position/position.component.ts
@@ -13,6 +13,11 @@ import { RouterLink, RouterModule, RouterOutlet } from '@angular/router';
 import { CreatepositionComponent } from '../createposition/createposition.component';
 import { ApiService } from '../../../../api.service';
 import { UpdatepositionComponent } from '../updateposition/updateposition.component';
+import { ButtonModule } from 'primeng/button';
+import { ConfirmDialog } from 'primeng/confirmdialog';
+import { RippleModule } from 'primeng/ripple';
+import { ToastModule } from 'primeng/toast';
+import { MessageService, ConfirmationService } from 'primeng/api';
 
 export interface position {
   id:number;
@@ -22,15 +27,18 @@ export interface position {
 
 @Component({
   selector: 'app-position',
-  imports: [MatFormFieldModule, MatInputModule, MatTableModule, MatButtonModule, RouterModule, ],
+  imports: [MatFormFieldModule, MatInputModule, MatTableModule, MatButtonModule, RouterModule,
+     ToastModule,ButtonModule,RippleModule,ConfirmDialog
+   ],
   templateUrl: './position.component.html',
-  styleUrl: './position.component.css'
+  styleUrl: './position.component.css',
+  providers: [MessageService,ConfirmationService]
 })
 export class PositionComponent implements OnInit{
  
   readonly dialog = inject(MatDialog);
-  readonly positionmanagementService = inject(ApiService);
-  
+  constructor(private messageService: MessageService,private positionmanagementService: ApiService, private confirmationService: ConfirmationService) {}
+
   dataSource = new MatTableDataSource<position>([]);
   displayedColumns: string[] = ['name', 'actions'];
 
@@ -58,12 +66,33 @@ export class PositionComponent implements OnInit{
     });
   }
 
-  deletepos(id: number): void {
-    this.positionmanagementService.deletepos(id).subscribe(() => {
-      this.dataSource.data = this.dataSource.data.filter(employee => employee.id !== id);
-    });
-  }
-
+  confirm(id: number) {
+    this.confirmationService.confirm({ 
+        header: 'Are you sure?',
+        message: 'Please confirm to proceed.',
+        accept: () => {
+          this.positionmanagementService.deletepos(id).subscribe(() => {
+            this.dataSource.data = this.dataSource.data.filter(employee => employee.id !== id);
+          
+                this.messageService.add({ 
+                    severity: 'success', 
+                    summary: 'Success', 
+                    detail: 'Deleted successfully',
+                    life: 3000
+                });
+               // Refresh the list after deletion
+            });
+        },
+        reject: () => {
+            this.messageService.add({ 
+                severity: 'info', 
+                summary: 'Cancel', 
+                detail: 'Deletion has been canceled' 
+            });
+        },
+        
+    }); this.getdata(); 
+}
   editposition(element: any) {
           const dialogRef = this.dialog.open(UpdatepositionComponent, {
             width: 'auto',
